:art: Typeset tables and figures in Chapter 2
Adjusting the placement and font of the figures to make it look nice. Needed to rotate one table with very long columns and shifted some paragraphs around. Made figures full \textwidth to fill up more of the page, and set positioning to "ht" or "htbp" instead of just "t". Also fixed the capitalization of one citation and italicized "The Cryosphere".